Use a react hook to access theme within linaria css-in-js
It seems there was some circular dependency that was causing `linaria`
to not transpile a file properly. Using a react hook resolves this.

It does mean that we can no longer access theme styles within external
`css` interpolations. But that seems like a step in the right direction
since it forces us to pass in that dependency explicitly.