Fix: the browser hanging while searching for constraint items
When a user types in a dropdown, the filtering logic caused the UI to
lock up. This degraded UX massively. The issue lied with the use of
fuzzy searching.

But fuzzy searching isn't really valuable for our case. Instead, a
contextual search is more appropriate. This commit uses `flexSearch` to
conduct contextual searches.

Closes: #76
Closes: #77

Squashed commits:
* Render a customized menu list for the select dropdown
* Render a virtualized dropdown list
* Keep the main popup open after selecting an item
* Scroll the item into view is the use scrolls past it with the keyboard
* Render the entire list of available item values
* Perf: improve searching in constraint inputs with flexSearch
* Increase the amount of filter results is returned