Create data visualizations for template queries
This commit adds the feature for generating visualizations for template
queries. There are now two available sections to run queries from:

1. Overview
2. Templates

Templates are pre-filtered by the selected class, and can be further
filtered by selecting a category. Each category displays the total count
of its templates. Note that some templates may be in more than one
category, so tallying the category counts may not match the total count.
The `Show All` category does display the correct total though.

Currently the constraints within the templates fetch their values each
time the popup is opened, and do not cache them when the components
unmount. The same is true for fetching templates. This will be addressed
in another story.

Closes: #101

Squashed commits:
* Allow switching between default and template views
* Rename "supervisor" to "appManager"
* Lift the default constraints into the app manager component
* Lift switching app views to the app manager component
* Display when no queries have been executed
* Render category tags
* Place category tags in an accordian
* Display just the "Run Query" button for templates
* Filter template constraints per class
* Filter templates by selected categories
* Remove categories with no available templates
* Display the count for each category
* Display only the tags for the selected class view
* Render checkbox items for ONE OF constraints
* Update the template query with the newly selected values
* Run a query and update visualizations
* Fetch values for a template constraint
* Display a tag above a constraint widget for every selected value
* Display the operation next to constraint widget tags
* Display success icon if the template query had been executed
* Display the description when hovering over the info icon
* Disable the QueryController for template view
* Display notification when no constraint items were received
* Extract categories to prevent Template component re-renders
* Update the categories when changing classes