Chore: create an overview machine with a config object
I previously had created a function to dynamically create overview
constraint machines. But this was a misguided premature abstraction, since
only the Overview component uses it. And xstate provides an idiomatic
way to pre-bind data to the context with the `machine.withContext` method.

This commit makes the overview machine a static configuration object, and
passes in each constraints required context data on creation from within
the component.

Closes: #106