Perf: Use xstate's assign rather than immer
`@xstate/immer` provides a nicer DX for updating the context. But it
fails miserably with large datasets. This is because it need to
recursively convert each item into an immutable draft. Considering that
intermine produces enormous amounts of data, and we don't modify that
data, we will not have an issue with mutations there.

The only place where mutation would cause problems is with our data
structures for state management. This is alleviated by using `xstate`'s
assign method.

Therefore, for performance reasons, this commit refactors to using
`xstate`s assign.

Closes: #108

Squashed commits:
* Refactor appManagerMachine to use xstate assign
* Refactor barChartMachine to use xstate assign
* Refactor overviewConstraintMachine to use xstate assign
* Refactor pieChartMachine to use xstate assign
* Revert "Refactor appManagerMachine to use xstate assign"
* This reverts commit b3d335557c1d7babc532c5ce14cc9c8d8eb5b6da.
* Refactor queryControllerMachine to use xstate assign
* Refactor tableMachine to use xstate assign
* Refactor templateConstraintMachine to use xstate assign
* Refactor templateQueryMachine to use xstate assign