Enable filtering queries with list constraints
This commit adds the ability to filter queries to the selected lists.
They can be selected from a filterable dropdown that also displays an
info button. This info button shows the description for the list when
hovering over it.

The selected lists are shown in different locations depending on the
query view:

1. overview constraints: shown alongside any other applied queries.
2. template constraints: shown above the category selection as tags

Users can also deselect lists from those locations.

Closes: #110

* Remove theme and reset buttons
* Move ClassSelector to its own file
* Fetch the lists for a mine
* Display a List selector widget
* Display the description when hovering over the info icon
* Construct a list config so that it can be filtered
* Display a title for the info icon popover
* Add list constraint to default view
* Add list constraint to the query controller constraints
* Remove list constraint from the query controller constraints
* Render list tags
* Remove list constraint from the template query