Show the status of whether lists are applied
At the moment the status of applying lists was only shown in both the
overview constraint `view all` popup, and the templates constraint `list`
section. This commit displays a tag next to the lists dropdown so that
the user doesn't have to open either of those to know if lists are applied.

Closes: #124