Save serialized search indexes in local storage
The indexes for filterable input widgets is expensive to build, and were
being rebuilt every time the components rendered. This commit save the
index to local storage so that they can be indexed just once.

N.B: This commit does not address setting an expiry date to this cache.
Further discussion needs to happen to decide on the best approach for that
moving forward.

Closes: #126