Fix: lists are not correctly formatted when being applied to queries
Currently the formatting for the lists has the list name placed in the
`values` prop of the query, but this keeps throwing errors. Instead, the
value should be placed in the `value` property since only 1 list can be
constrained per query.

This commit resolves multiple issues:

1. the lists not being properly formatted
2. allowing only 1 list to be selected
3. adding the list to the overview constraint query

Because we now only show 1 list, the lists have been removed from the
overview and templates sections. Instead, the selected value is placed
in a tag next to the dropdown widget. It is also placed above the other
values within the dropdown, where it can be unselected by clicking on it.

Closes: #130
Closes: #131
Closes: #132

Squashed commits:
* Place the selected list at the top of the menu
* Remove the selected list when it is clicked again
* Run an overview query constraint on a list
* Run an template query constrained on a list
* Truncate the text inside a constraint tag