Render overview constraints with spawned actors
Because we no longer were updating the overview constraints with an event
bus action, they would not refresh when the mine or class view changed.
This commit renders the overview constraints as spawned actors whenever
the mine or class changes.

This commit also ensures that previous spawned actors are stopped, so that
they can no longer receive events and transition.

Closes: #140