Show default values for template constraints
This commit displays the default values for template constraints, along
with the operation used for it. If a constraint is completely removed,
the `Run Query` button will be disabled, so that the templates stay
properly formatted. There is also an option to reset all constraints to
the default values.

For now, constraints that have more than 1000 items are displayed as an
Input widget. This is for 2 reasons:

1. Select dropdowns are not virtualized, and rendering that many elements
will cause the UI to seriously lag.

2. Bluegenes currently does something similar. Although the actual cutoff
may not be the same as this app.

Closes: #147

Squashed commits:
* Only fetch constraint values if it has any
* Render a SelectWidget for single selection constraints
* Display default constraint selections
* Display a dropdown for single select constraints
* Disallow duplicate value selections
* Display an operation label for template constraints
* Allow resetting constraints to the default value
* Disable run query button if a constraint has no value
* Render Input widgets