Show loaders while the mine configuration loads
When first loading the app, the data viz shows no results are available,
which leads to the belief there is an error. And when the mine changes,
the data viz does not update, making it seem that the UI has hung up.

This commit resolves both issues by showing loading skeleton screens at
first load, and whenever the mine changes.

Closes: #145

Squashed commits:
* Move mine/class change logic to machine
* Show loaders when mine changes
* Show loaders while the mine loads