Fix: decrease memory consumption
The app was taking 500-600mb of memory according to chrome's task
manager, and the heap was reading close to 200mb. This happened because
of the size of the search indexes. Since the search is only used for direct
matches, decreasing the encoder specificity and the tokenizer to index only
forward decreased the heap to 70mb.

Closes: #176