Fix: navigation buttons squish when shrinking the page
When shrinking the page, the navigation buttons shrink. Although full
mobile responsiveness is being pushed into a future story, because it will
require a mobile layout design this is still not ideal even for desktops.
Now the navigation bar and the table are mobile responsive down to about
iPad pro tablet sizes.

Closes: #184

Squashed commits:
ˆ* make navbar responsive
ˆ* make data viz section mobile responsive to ipad pro size