Open the generate code dialog when a new language is selected
When the user selects a new language to generate code, they then had to
also click on the language button. Doing this twice to open the dialog
is not a nice experience. This commit changes the language and opens the
dialog should they change the language from the table view.

Closes: #186