Display a notification if the class does not have a path available
Some classes do not have a path available, yet are displaying a warning
when trying to process data for the charts. This is not ideal since these
classes not having those paths is a known state, and not an error.

This commit displays a standard notification stating the correct state
of the application.

Closes: #196

Squashed commits:
diplay an informative component for the pie chart
display an informative component for the bar chart